+++ NEWS.md
@@ -1,3 +1,4 @@
+* added focus module 20200331 epidemiologia comparativa Veneto-Piemonte
 * removed unuseful button from plotly
 * added istructions to manage plots
 * colour and label name for levels (#33)